full split into several *.h files
